+++ cstree/src/interning/lasso_compat/token_interner.rs
@@ -87,10 +87,12 @@ pub use multi_threaded::MultiThreadedTokenInterner;
 mod multi_threaded {
     use super::*;
 
+    use std::sync::Arc as StdArc;
+
     /// A threadsafe [`Interner`] for deduplicating [`GreenToken`](crate::green::GreenToken) strings.
     ///
-    /// Note that [`Interner`] and [`Resolver`] are also implemented for  `&MultiThreadTokenInterner` so you can pass
-    /// `&mut &interner` in shared contexts.
+    /// Note that [`Interner`] and [`Resolver`] are also implemented for `&MultiThreadTokenInterner` and
+    /// `Arc<MultiThreadTokenInterner>` so you can pass a mutable reference to either of these in shared contexts.
     #[cfg_attr(doc_cfg, doc(cfg(feature = "multi_threaded_interning")))]
     pub struct MultiThreadedTokenInterner {
         rodeo: ThreadedRodeo<TokenKey, FxBuildHasher>,
@@ -114,6 +116,5 @@ mod multi_threaded {
     }
 
     impl_traits!(for MultiThreadedTokenInterner, if #[cfg(feature = "multi_threaded_interning")]);
-
-    impl_traits!(for &MultiThreadedTokenInterner, if #[cfg(feature = "multi_threaded_interning")]);
+    impl_traits!(for StdArc<MultiThreadedTokenInterner>, if #[cfg(feature = "multi_threaded_interning")]);
 }